requirements.txt
asgiref==3.7.2
Django==4.2.6
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
EasyProcess==1.1
entrypoint2==1.1
inflection==0.5.1
MouseInfo==0.1.3
mss==9.0.1
packaging==23.2
Pillow==10.0.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-decouple==3.8
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
